+++ src/main/java/roomescape/admin/controller/AdminController.java
@@ -1,0 +1,28 @@
+package roomescape.admin.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AdminController {
+
+    @GetMapping("/admin")
+    public String showAdminPage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String showAdminReservationPage() {
+        return "admin/reservation-new";
+    }
+
+    @GetMapping("/admin/time")
+    public String showAdminTimePage() {
+        return "admin/time";
+    }
+
+    @GetMapping("/admin/theme")
+    public String showAdminThemePage() {
+        return "admin/theme";
+    }
+}
