+++ src/test/java/roomescape/member/dao/MemberDaoTest.java
@@ -1,0 +1,8 @@
+package roomescape.member.dao;
+
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+
+@JdbcTest
+class MemberDaoTest {
+
+}
